requirements.txt
alembic==1.11.1
annotated-types==0.7.0
anyio==3.7.1
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.21.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-ai-inference==1.0.0b9
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.100.1
fastapi-pagination==0.12.17
greenlet==3.2.2
gunicorn==20.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
msal==1.32.0
msal-extensions==1.3.1
numpy==2.2.6
openai==1.75.0
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.18
uvicorn==0.23.1
